backend/pyproject.toml
[tool.poetry]
name = "clipmind-backend"
version = "0.1.0"
description = "ClipMind AI Video Search Backend"
authors = ["Abhishek Kankanala <abhikank90@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
# FastAPI and server
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"

# Database
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
psycopg2-binary = "^2.9.9"

# Pydantic
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

# Celery and Redis
celery = "^5.3.4"
redis = "^5.0.1"

# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

# AWS
boto3 = "^1.34.0"

# AI and ML
torch = "^2.1.0"
torchvision = "^0.16.0"
transformers = "^4.35.0"
sentence-transformers = "^2.2.2"
openai-whisper = "^20231117"

# Video Processing
opencv-python = "^4.8.1"
Pillow = "^10.1.0"
ffmpeg-python = "^0.2.0"
scenedetect = {extras = ["opencv"], version = "^0.6.2"}

# Vector Database
pinecone-client = "^2.2.4"

# Utilities
numpy = "^1.24.3"
httpx = "^0.25.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
